// interfaces/homeProducts.tsx
// Generated by https://quicktype.io

export interface Products {
    id:            string;
    storeId:       string;
    categoryID:    string;
    subcategoryID: string;
    hasPrice:      boolean;
    isFeatured:    boolean;
    isArchived:    boolean;
    price:         number;
    name:          string;
    description:   string;
    createdAt:     string;
    updatedAt:     string;
    images:        Image[];
    category:      Category;
    subcategory:   Subcategory;
    quantity?: number;
}

export interface Category {
    id:          string;
    storeId:     string;
    billboardID: string;
    name:        string;
    createdAt:   string;
    updatedAt:   string;
}

export interface Image {
    id:        string;
    productID: string;
    url:       string;
    createdAt: string;
    updatedAt: string;
}

export interface Subcategory {
    id:         string;
    storeId:    string;
    categoryID: string;
    hasPrice:   boolean;
    name:       string;
    value:      string;
    createdAt:  string;
    updatedAt:  string;
}